requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Reuploaded==1.2.0
python-dotenv==1.0.0
Pillow==10.0.0
email-validator==2.0.0
Werkzeug==2.2.3
